create zip for download

+++ putio.py
@@ -3,18 +3,27 @@ import os
 import re
 import json
 import logging
+import time
 import webbrowser
 from urllib import urlencode
+from datetime import datetime
 
 import requests
-import iso8601
+from requests.adapters import HTTPAdapter
+from requests.packages.urllib3.util.retry import Retry
+
+KB = 1024
+MB = 1024 * KB
+
+# Read and write operations are limited to this chunk size.
+# This can make a big difference when dealing with large files.
+CHUNK_SIZE = 256 * KB
 
 BASE_URL = 'https://api.put.io/v2'
 ACCESS_TOKEN_URL = 'https://api.put.io/v2/oauth2/access_token'
 AUTHENTICATION_URL = 'https://api.put.io/v2/oauth2/authenticate'
 
 logger = logging.getLogger(__name__)
-
 
 class AuthHelper(object):
 
@@ -53,9 +62,20 @@ class AuthHelper(object):
 
 class Client(object):
 
-    def __init__(self, access_token):
+    def __init__(self, access_token, use_retry=False):
         self.access_token = access_token
         self.session = requests.session()
+
+        if use_retry:
+            # Retry maximum 10 times, backoff on each retry
+            # Sleeps 1s, 2s, 4s, 8s, etc to a maximum of 120s between retries
+            # Retries on HTTP status codes 500, 502, 503, 504
+            retries = Retry(total=10,
+                            backoff_factor=1,
+                            status_forcelist=[500, 502, 503, 504])
+
+            # Use the retry strategy for all HTTPS requests
+            self.session.mount('https://', HTTPAdapter(max_retries=retries))
 
         # Keep resource classes as attributes of client.
         # Pass client to resource classes so resource object
@@ -63,6 +83,7 @@ class Client(object):
         attributes = {'client': self}
         self.File = type('File', (_File,), attributes)
         self.Transfer = type('Transfer', (_Transfer,), attributes)
+        self.Account = type('Account', (_Account,), attributes)
 
     def request(self, path, method='GET', params=None, data=None, files=None,
                 headers=None, raw=False, stream=False):
@@ -119,8 +140,8 @@ class _BaseResource(object):
         self.name = None
         self.__dict__.update(resource_dict)
         try:
-            self.created_at = iso8601.parse_date(self.created_at)
-        except (AttributeError, iso8601.ParseError):
+            self.created_at = strptime(self.CReated_at)
+        except Exception:
             self.created_at = None
 
     def __str__(self):
@@ -148,28 +169,50 @@ class _File(_BaseResource):
         return [cls(f) for f in files]
 
     @classmethod
-    def upload(cls, path, name=None):
+    def upload(cls, path, name=None, parent_id=0):
         with open(path) as f:
             if name:
                 files = {'file': (name, f)}
             else:
                 files = {'file': f}
-            d = cls.client.request('/files/upload', method='POST', files=files)
-
+            d = cls.client.request('/files/upload', method='POST',
+                                   data={'parent_id': parent_id}, files=files)
+        logger.info(d)
         f = d['file']
         return cls(f)
+
+    @classmethod
+    def createZip(cls,file_ids):
+        d = cls.client.request('/files/zip',
+                                method='GET',params={'file_ids': file_ids})
+        if d['status'] == "OK":
+            zip_id = d['zip_id']
+            notReady = True
+	    while notReady:
+	        time.sleep(5)
+                d = cls.client.request('/zips/%i' % zip_id, method='GET')
+                #logger.info(d)
+                if d['url'] != False:
+                    notReady= False
+            return d['url']
+        else:
+            return None
+
+    @classmethod
+    def zipInfo(cls,zip_id):
+        return cls.client.request('/zips/%i' % zip_id, method='GET')
 
     def dir(self):
         """List the files under directory."""
         return self.list(parent_id=self.id)
 
-    def download(self, dest='.', delete_after_download=False):
+    def download(self, dest='.', delete_after_download=False, chunk_size=CHUNK_SIZE):
         if self.content_type == 'application/x-directory':
-            self._download_directory(dest, delete_after_download)
+            self._download_directory(dest, delete_after_download, chunk_size)
         else:
-            self._download_file(dest, delete_after_download)
+            self._download_file(dest, delete_after_download, chunk_size)
 
-    def _download_directory(self, dest='.', delete_after_download=False):
+    def _download_directory(self, dest, delete_after_download, chunk_size):
         name = self.name
         if isinstance(name, unicode):
             name = name.encode('utf-8', 'replace')
@@ -179,33 +222,91 @@ class _File(_BaseResource):
             os.mkdir(dest)
 
         for sub_file in self.dir():
-            sub_file.download(dest, delete_after_download)
+            sub_file.download(dest, delete_after_download, chunk_size)
 
         if delete_after_download:
             self.delete()
 
-    def _download_file(self, dest='.', delete_after_download=False):
-        response = self.client.request(
-            '/files/%s/download' % self.id, raw=True, stream=True)
+    def _verify_file(self, filepath):
+        logger.info('verifying crc32...')
+        filesize = os.path.getsize(filepath)
+        if self.size != filesize:
+            logging.error('file %s is %d bytes, should be %s bytes' % (filepath, filesize, self.size))
+            return False
 
-        filename = re.match(
-            'attachment; filename=(.*)',
-            response.headers['content-disposition']).groups()[0]
-        # If file name has spaces, it must have quotes around.
-        filename = filename.strip('"')
+        crcbin = 0
+        with open(filepath, 'rb') as f:
+            while True:
+                chunk = f.read(CHUNK_SIZE)
+                if not chunk:
+                    break
 
-        with open(os.path.join(dest, filename), 'wb') as f:
-            for chunk in response.iter_content(chunk_size=1024):
-                if chunk:  # filter out keep-alive new chunks
-                    f.write(chunk)
-                    f.flush()
+                crcbin = binascii.crc32(chunk, crcbin) & 0xffffffff
 
-        if delete_after_download:
-            self.delete()
+        crc32 = '%08x' % crcbin
+
+        if crc32 != self.crc32:
+            logging.error('file %s CRC32 is %s, should be %s' % (filepath, crc32, self.crc32))
+            return False
+
+        return True
+
+    def _download_file(self, dest, delete_after_download, chunk_size):
+        name = self.name
+        if isinstance(name, unicode):
+            name = name.encode('utf-8', 'replace')
+
+        filepath = os.path.join(dest, name)
+        if os.path.exists(filepath):
+            first_byte = os.path.getsize(filepath)
+
+            if first_byte == self.size:
+                logger.warning('file %s exists and is the correct size %d' % (filepath, self.size))
+        else:
+            first_byte = 0
+
+        logger.debug('file %s is currently %d, should be %d' % (filepath, first_byte, self.size))
+
+        if self.size == 0:
+            # Create an empty file
+            open(filepath, 'w').close()
+            logger.debug('created empty file %s' % filepath)
+        else:
+            if first_byte < self.size:
+                with open(filepath, 'ab') as f:
+                    headers = {'Range': 'bytes=%d-' % first_byte}
+
+                    logger.debug('request range: bytes=%d-' % first_byte)
+                    response = self.client.request('/files/%s/download' % self.id,
+                                                headers=headers,
+                                                raw=True,
+                                                stream=True)
+
+                    for chunk in response.iter_content(chunk_size=chunk_size):
+                        if chunk:  # filter out keep-alive new chunks
+                            f.write(chunk)
+
+        if self._verify_file(filepath):
+            if delete_after_download:
+                self.delete()
 
     def delete(self):
         return self.client.request('/files/delete', method='POST',
-                                   data={'file_ids': str(self.id)})
+                                   data={'file_id': str(self.id)})
+
+    @classmethod
+    def delete_multi(cls, ids):
+        return cls.client.request('/files/delete', method='POST',
+                                  data={'file_ids': ','.join(map(str, ids))})
+
+    def move(self, parent_id):
+        return self.client.request('/files/move', method='POST',
+                                   data={'file_ids': str(self.id), 'parent_id': str(parent_id)})
+
+    def rename(self, name):
+        return self.client.request('/files/rename', method='POST',
+                                   data={'file_id': str(self.id), 'name': str(name)})
+
 
 
 class _Transfer(_BaseResource):
@@ -225,14 +326,14 @@ class _Transfer(_BaseResource):
     @classmethod
     def add_url(cls, url, parent_id=0, extract=False, callback_url=None):
         d = cls.client.request('/transfers/add', method='POST', data=dict(
-            url=url, parent_id=parent_id, extract=extract,
+            url=url, save_parent_id=parent_id, extract=extract,
             callback_url=callback_url))
         t = d['transfer']
         return cls(t)
 
     @classmethod
     def add_torrent(cls, path, parent_id=0, extract=False, callback_url=None):
-        with open(path) as f:
+        with open(path, 'rb') as f:
             files = {'file': f}
             d = cls.client.request('/files/upload', method='POST', files=files,
                                    data=dict(parent_id=parent_id,
@@ -244,3 +345,43 @@ class _Transfer(_BaseResource):
     @classmethod
     def clean(cls):
         return cls.client.request('/transfers/clean', method='POST')
+
+    def cancel(self):
+        return self.client.request('/transfers/cancel',
+                                   method='POST',
+                                   data={'transfer_ids': self.id})
+
+    @classmethod
+    def cancel_multi(cls, ids):
+        return cls.client.request('/transfers/cancel',
+                                  method='POST',
+                                  data={'transfer_ids': ','.join(map(str, ids))})
+
+
+class _Account(_BaseResource):
+
+    @classmethod
+    def info(cls):
+        return cls.client.request('/account/info', method='GET')
+
+    @classmethod
+    def settings(cls):
+        return cls.client.request('/account/settings', method='GET')
+
+
+# Due to a nasty bug in datetime module, datetime.strptime calls
+# are not thread-safe and can throw a TypeError. Details: https://bugs.python.org/issue7980
+# Here we are implementing simple RFC3339 parser which is used in Put.io APIv2.
+def strptime(date):
+    """Returns datetime object from the given date, which is in a specific format: YYYY-MM-ddTHH:mm:ss"""
+    d = {
+            'year': date[0:4],
+            'month': date[5:7],
+            'day': date[8:10],
+            'hour': date[11:13],
+            'minute': date[14:16],
+            'second': date[17:],
+            }
+
+    d = dict((k, int(v)) for k, v in d.iteritems())
+    return datETime(**d)
